refactor(stat): apply JPA to stat repository

+++ src/main/kotlin/ecommerce/dto/ProductStatsResponse.kt
@@ -4,6 +4,6 @@ import java.time.LocalDateTime
 
 class ProductStatsResponse(
     val productName: String,
-    val productQuantity: Long,
+    val productQuantity: Int,
     val mostRecent: LocalDateTime,
 )
